square is typedef'ed as an int, so use that for output.

+++ debugging/trace.c
@@ -249,7 +249,7 @@ void TraceSquareImpl(char const *prefix, square s)
       if (is_on_board(s))
         WriteSquare(&output_plaintext_engine,stdout,s);
       else
-        output_plaintext_engine.fprintf(stdout, "[square %u]", (unsigned int)s);
+        output_plaintext_engine.fprintf(stdout, "[square %d]", s);
     }
     fflush(stdout);
   }
